kvstreamer: move the package from kv/kvclient into sql
Seems like `pkg/sql` is a better place for the `kvstreamer` package
since SQL Queries is the owner of it, it now has a dependency on
`sql/rowcontainer`, and is mainly used by `sql/row`.

Release note: None

Release justification: bug fixes and low-risk updates to new functionality.
