+++ pkg/sql/colfetcher/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "//pkg/col/typeconv",
         "//pkg/keys",
         "//pkg/kv",
-        "//pkg/kv/kvclient/kvstreamer",
         "//pkg/roachpb",
         "//pkg/sql/catalog",
         "//pkg/sql/catalog/catpb",
@@ -32,6 +31,7 @@ go_library(
         "//pkg/sql/colmem",
         "//pkg/sql/execinfra",
         "//pkg/sql/execinfrapb",
+        "//pkg/sql/kvstreamer",
         "//pkg/sql/memsize",
         "//pkg/sql/row",
         "//pkg/sql/rowenc",
