+++ pkg/sql/row/BUILD.bazel
@@ -29,7 +29,6 @@ go_library(
         "//pkg/jobs/jobspb",
         "//pkg/keys",
         "//pkg/kv",
-        "//pkg/kv/kvclient/kvstreamer",
         "//pkg/kv/kvserver",
         "//pkg/kv/kvserver/concurrency/lock",
         "//pkg/roachpb",
@@ -43,6 +42,7 @@ go_library(
         "//pkg/sql/catalog/descs",
         "//pkg/sql/catalog/schemaexpr",
         "//pkg/sql/catalog/seqexpr",
+        "//pkg/sql/kvstreamer",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/rowenc",
