+++ pkg/sql/rowexec/BUILD.bazel
@@ -44,7 +44,6 @@ go_library(
         "//pkg/jobs/jobspb",
         "//pkg/keys",
         "//pkg/kv",
-        "//pkg/kv/kvclient/kvstreamer",
         "//pkg/kv/kvserver/concurrency/lock",
         "//pkg/kv/kvserver/kvserverbase",
         "//pkg/roachpb",
@@ -59,6 +58,7 @@ go_library(
         "//pkg/sql/execinfra",
         "//pkg/sql/execinfrapb",
         "//pkg/sql/inverted",
+        "//pkg/sql/kvstreamer",
         "//pkg/sql/memsize",
         "//pkg/sql/opt/invertedexpr",
         "//pkg/sql/opt/invertedidx",
